Make xml references syntactically correct
We had several references in xml summaries that were incorrect.
Most of them either referenced long gone classes or those classes are now
generic.

Fixes methods were only some parameters were documented.

+++ Tests/FluentAssertions.Specs/Equivalency/CollectionEquivalencySpecs.cs
@@ -1361,7 +1361,7 @@ namespace FluentAssertions.Specs
 
             // Act
             Action act = () => subject.Should().BeEquivalentTo(expectation, opts => opts
-                .Using<int>(ctx => ((int)ctx.Subject).Should().BeInRange(ctx.Expectation - 1, ctx.Expectation + 1))
+                .Using<int>(ctx => ctx.Subject.Should().BeInRange(ctx.Expectation - 1, ctx.Expectation + 1))
                 .WhenTypeIs<int>()
             );
